requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.13.2
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-core==1.2.20
langchain-google-genai==2.0.10
langchain-groq==1.1.2
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.0.2
openai==2.20.0
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.46
stack-data==0.6.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.2
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.6.0
wrapt==2.1.2
zstandard==0.25.0
